requirements.txt
cogapp==3.0.0
dotmap==1.3.14
PyYAML==5.3.1
prompt-toolkit==1.0.14
Pygments==2.6.1
PyInquirer==1.0.3
regex==2020.6.8
six==1.15.0
wcwidth==0.2.4